big fixed
when click enter on the input, there are two things happen. when the result is showing and one of item in the list has been selected, value will be replaced with selected data and won't submit for a form. when the result is showing but nothing been selected enter the input will trigger form submit if any